requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-google-genai
langchain-deepseek
pandas
mistralai

